o  ID (long) related updates to Store interface, impl and tests.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497753 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-api/src/main/java/org/apache/maven/continuum/Continuum.java
@@ -87,7 +87,7 @@ public interface Continuum
     // Project
     // ----------------------------------------------------------------------
 
-    void removeProject( int projectId )
+    void removeProject( long projectId )
         throws ContinuumException;
 
     void checkoutProject( int projectId )
